+++ CHANGES.md
@@ -68,6 +68,10 @@ To be released.
 
  -  Implemented `list()` method in `RedisKvStore`.  [[#498]]
 
+### @fedify/denokv
+
+ -  Implemented `list()` method in `DenoKvStore`.  [[#498]]
+
 
 Version 1.9.2
 -------------